requirements.txt
Flask
tensorflow
ultralytics
opencv-python
supervision
Pillow
numpy
